refactor: типы данных и расположение файлов
