+++ src/components/model/index.ts
