+++ src/components/service/shop-api.ts
@@ -3,9 +3,9 @@ import {
   IOrder,
   IProduct,
   IShopAPI,
-  OrderResponseSuccess,
-} from '../types';
-import { Api } from '../components/base/api';
+  IOrderResult,
+} from '../../types';
+import { Api } from '../base/api';
 
 export class ShopAPI extends Api implements IShopAPI {
   async getProducts(): Promise<ApiListResponse<IProduct>> {
@@ -13,10 +13,10 @@ export class ShopAPI extends Api implements IShopAPI {
   }
 
   async getProductById(id: string): Promise<IProduct> {
-    return (await this.get(`/product.${id}`)) as Promise<IProduct>;
+    return (await this.get(`/product/${id}`)) as Promise<IProduct>;
   }
 
-  async createOrder(order: IOrder): Promise<OrderResponseSuccess> {
-    return (await this.post('/order', order)) as Promise<OrderResponseSuccess>;
+  async createOrder(order: IOrder): Promise<IOrderResult> {
+    return (await this.post('/order', order)) as Promise<IOrderResult>;
   }
 }
